refactor: refine tx history types (#4402)
