+++ src/models/transaction.ts
@@ -1,4 +1,4 @@
-import { BigNumberish, Signer, Transaction, providers } from 'ethers'
+import { BigNumberish, Signer, Transaction } from 'ethers'
 
 export enum TxStatus {
   pending = 'PENDING',
@@ -20,20 +20,3 @@ export interface TransactionOptions extends TransactionCallbacks {
   value?: BigNumberish
 }
 
-export type TransactionLog = {
-  id: number
-  title: string
-  createdAt: number
-  callbacks?: TransactionCallbacks
-} & (
-  | {
-      // Only pending txs have not been mined
-      status: TxStatus.pending
-      tx: Transaction | null
-    }
-  | {
-      // Once mined, tx will be a TransactionResponse
-      status: TxStatus.success | TxStatus.failed
-      tx: providers.TransactionResponse | null
-    }
-)
